LowLatencyLogger mods

+++ SdFat/examples/LowLatencyLogger/UserTypes.h
@@ -5,8 +5,8 @@
 #define FILE_BASE_NAME "adc4pin"
 const uint8_t ADC_DIM = 4;
 struct data_t {
-  unsigned long time;
-  unsigned short adc[ADC_DIM];
+  uint32_t time;
+  uint16_t adc[ADC_DIM];
 };
 void acquireData(data_t* data);
 void printData(Print* pr, data_t* data);
